extract function
