+++ .github/workflows/e2e.yml
@@ -141,7 +141,7 @@ jobs:
       - run: unzip ${{ env.FILE_NAME }} -d e2e-tests
       
       - name: run e2e tests
-        run: node e2e/out/runFtTest $(pwd)/e2e-tests/extension
+        run: node e2e/out/runners/runFtTest $(pwd)/e2e-tests/extension
 
       - name: upload logs
         if: always()
